+++ test/test_alm.jl
@@ -54,3 +54,16 @@ alm = Healpix.readAlmFromFITS("alm.fits", ComplexF64)
 ## test alm2cl
 testalm = Healpix.Alm(2, 2, ComplexF64.(1:6))
 @test isapprox(Healpix.alm2cl(testalm), [1.0, 12.0, 26.2])
+
+
+## test gaussbeam
+
+# spin-0
+bl_0 = Healpix.gaussbeam(deg2rad(1), 512; pol=false)
+@test bl_0[50+1] ≈ 0.9323560235080715
+@test bl_0[500+1] ≈ 0.0010276785784086898
+
+# spin-2
+bl_2 = Healpix.gaussbeam(deg2rad(1), 512; pol=true)
+@test bl_2[50+1] ≈ 0.9324584647703964
+@test bl_2[500+1] ≈ 0.001027791493098124