+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+- New function `gaussbeam` [#85](https://github.com/ziotom78/Healpix.jl/pull/85)
+
 # Version 4.0.1
 
 -   Add checks for colatitude θ in functions that take angles as input [#84](https://github.com/ziotom78/Healpix.jl/pull/84)